+++ .github/workflows/test-k2-as-third-party-lib-cpu-macos.yml
@@ -1,4 +1,4 @@
-# Copyright      2021  Fangjun Kuang (csukuangfj@gmail.com)
+# Copyright      2022  Fangjun Kuang (csukuangfj@gmail.com)
 
 # See ../../LICENSE for clarification regarding multiple authors
 #
@@ -16,7 +16,7 @@
 
 # refer to https://github.com/actions/starter-workflows/pull/47/files
 
-name: build-cpu-macos
+name: test-k2-as-third-party-lib-cpu-macos
 
 on:
   push:
@@ -42,11 +42,11 @@ jobs:
         id: set-matrix
         run: |
           # outputting for debugging purposes
-          python scripts/github_actions/generate_build_matrix.py
-          MATRIX=$(python scripts/github_actions/generate_build_matrix.py)
+          python scripts/github_actions/generate_build_matrix.py --test-only-latest-torch
+          MATRIX=$(python scripts/github_actions/generate_build_matrix.py --test-only-latest-torch)
           echo "::set-output name=matrix::${MATRIX}"
 
-  build-cpu-macos:
+  test-k2-as-third-party-lib-cpu-macos:
     if: github.event.label.name == 'ready' || github.event_name == 'push'
     needs: generate_build_matrix
     runs-on: macos-10.15
@@ -85,39 +85,24 @@ jobs:
         run: |
           python3 -m pip install -qq --upgrade pip
           python3 -m pip install -qq wheel twine dataclasses
-          python3 -m pip install -qq torch==${{ matrix.torch }}
+          python3 -m pip install -qq torch==${{ matrix.torch }} numpy
 
           python3 -c "import torch; print('torch version:', torch.__version__)"
-
-      - name: Build k2
-        shell: bash
-        run: |
-          pwd
-          mkdir build
-          cd build
-          cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DK2_WITH_CUDA=OFF ..
-          cat k2/csrc/version.h
-          cat CMakeCache.txt
-
-          make VERBOSE=1 -j2
-
-      - name: Run tests
-        shell: bash
-        run: |
-          cd build
-          ctest --output-on-failure
 
       - name: Build wheel
         shell: bash
         run: |
           export K2_CMAKE_ARGS="-DCMAKE_BUILD_TYPE=$BUILD_TYPE -DK2_WITH_CUDA=OFF"
           export K2_MAKE_ARGS="-j2"
-          python3 setup.py bdist_wheel
-          ls -lh dist/
-          ls -lh build/*
+          python3 setup.py install
+          python3 -m k2.version
+          python3 setup.py clean
+          python3 -m k2.version
 
-      - name: Upload Wheel
-        uses: actions/upload-artifact@v2
-        with:
-          name: torch-${{ matrix.torch }}-python-${{ matrix.python-version }}-macos-10.15-cpu
-          path: dist/*.whl
+      - name: test k2
+        shell: bash
+        run: |
+          cat $(python3 -c "import k2; print(k2.cmake_prefix_path)")/k2/k2Config.cmake
+          cd ./scripts/github_actions/k2-torch-api-test
+          make
+
