+++ .github/workflows/test-k2-as-third-party-lib-cpu-windows.yml
@@ -1,4 +1,4 @@
-# Copyright      2021  Fangjun Kuang (csukuangfj@gmail.com)
+# Copyright      2022  Fangjun Kuang (csukuangfj@gmail.com)
 
 # See ../../LICENSE for clarification regarding multiple authors
 #
@@ -16,7 +16,7 @@
 
 # refer to https://github.com/actions/starter-workflows/pull/47/files
 
-name: build-cpu-macos
+name: test-k2-as-third-party-lib-cpu-windows
 
 on:
   push:
@@ -42,14 +42,14 @@ jobs:
         id: set-matrix
         run: |
           # outputting for debugging purposes
-          python scripts/github_actions/generate_build_matrix.py
-          MATRIX=$(python scripts/github_actions/generate_build_matrix.py)
+          python scripts/github_actions/generate_build_matrix.py --test-only-latest-torch
+          MATRIX=$(python scripts/github_actions/generate_build_matrix.py --test-only-latest-torch)
           echo "::set-output name=matrix::${MATRIX}"
 
-  build-cpu-macos:
+  test-k2-as-third-party-lib-cpu-windows:
     if: github.event.label.name == 'ready' || github.event_name == 'push'
     needs: generate_build_matrix
-    runs-on: macos-10.15
+    runs-on: windows-2019
     strategy:
       fail-fast: false
       matrix:
@@ -61,16 +61,9 @@ jobs:
         with:
           fetch-depth: 0
 
-      - uses: szenius/set-timezone@v1.0
-        with:
-          timezoneLinux: "Asia/Shanghai"
-
-      - name: Display date and time
-        run: date
-
-      - name: Display clang version
-        run: |
-          clang --version
+      # see https://github.com/microsoft/setup-msbuild
+      - name: Add msbuild to PATH
+        uses: microsoft/setup-msbuild@v1.0.2
 
       - name: Setup Python ${{ matrix.python-version }}
         uses: actions/setup-python@v2
@@ -81,43 +74,28 @@ jobs:
         run: python -c "import sys; print(sys.version)"
 
       - name: Install PyTorch ${{ matrix.torch }}
+        run: |
+          pip3 install -qq torch==${{ matrix.torch }}+cpu -f https://download.pytorch.org/whl/torch_stable.html numpy
+          pip3 install -qq wheel twine dataclasses typing_extensions
+
+          python3 -m torch.utils.collect_env
+
+      - name: Install k2
         shell: bash
         run: |
-          python3 -m pip install -qq --upgrade pip
-          python3 -m pip install -qq wheel twine dataclasses
-          python3 -m pip install -qq torch==${{ matrix.torch }}
+          python3 setup.py install
+          python3 -m k2.version
+          python3 setup.py clean
+          python3 -m k2.version
 
-          python3 -c "import torch; print('torch version:', torch.__version__)"
-
-      - name: Build k2
+      - name: test k2
         shell: bash
         run: |
-          pwd
+          cat $(python3 -c "import k2; print(k2.cmake_prefix_path)")/k2/k2Config.cmake
+          cd ./scripts/github_actions/k2-torch-api-test
           mkdir build
           cd build
-          cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DK2_WITH_CUDA=OFF ..
-          cat k2/csrc/version.h
-          cat CMakeCache.txt
-
-          make VERBOSE=1 -j2
-
-      - name: Run tests
-        shell: bash
-        run: |
-          cd build
-          ctest --output-on-failure
-
-      - name: Build wheel
-        shell: bash
-        run: |
-          export K2_CMAKE_ARGS="-DCMAKE_BUILD_TYPE=$BUILD_TYPE -DK2_WITH_CUDA=OFF"
-          export K2_MAKE_ARGS="-j2"
-          python3 setup.py bdist_wheel
-          ls -lh dist/
-          ls -lh build/*
-
-      - name: Upload Wheel
-        uses: actions/upload-artifact@v2
-        with:
-          name: torch-${{ matrix.torch }}-python-${{ matrix.python-version }}-macos-10.15-cpu
-          path: dist/*.whl
+          ls -l
+          cmake ..
+          cmake --build . --target ALL_BUILD --config Release
+          ctest -C Release --output-on-failure
