+++ .gitignore
@@ -1,3 +1,4 @@
+torch_version.py
 # Build folder
 **/build*
