+++ get_version.py
@@ -41,7 +41,7 @@ def get_cuda_version():
     running_cuda_version = collect_env.get_running_cuda_version(
         collect_env.run)
     cuda_version = torch.version.cuda
-    if running_cuda_version is not None:
+    if running_cuda_version is not None and cuda_version is not None:
         assert cuda_version in running_cuda_version, \
                 f'PyTorch is built with CUDA version: {cuda_version}.\n' \
                 f'The current running CUDA version is: {running_cuda_version}'
