+++ scripts/github_actions/install_cudnn.sh
@@ -47,7 +47,7 @@ esac
 
 command -v git-lfs >/dev/null 2>&1 || { echo >&2 "\nPlease install 'git-lfs' first."; exit 2; }
 
-git clone https://huggingface.co/csukuangfj/cudnn
+GIT_LFS_SKIP_SMUDGE=1 git clone https://huggingface.co/csukuangfj/cudnn
 cd cudnn
 git lfs pull --include="$filename"
