+++ scripts/github_actions/k2-torch-api-test/CMakeLists.txt
@@ -13,6 +13,7 @@ include(pybind11)
 include(torch)
 include(k2)
 
+enable_testing()
 include(googletest)
 
 add_executable(torch_api_test torch_api_test.cc)
@@ -22,4 +23,15 @@ target_link_libraries(torch_api_test
       gtest_main
       ${TORCH_LIBRARIES}
       ${K2_LIBRARIES}
+)
+
+# NOTE: We set the working directory here so that
+# it works also on windows. The reason is that
+# the required DLLs are inside ${TORCH_DIR}/lib
+# and they can be found by the exe if the current
+# working directory is ${TORCH_DIR}\lib
+add_test(NAME "Test.torch_api_test"
+  COMMAND
+  $<TARGET_FILE:torch_api_test>
+  WORKING_DIRECTORY ${TORCH_DIR}/lib
 )
